+++ src/scalars/components/examples/enum-field-example/enum-field-example.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@powerhousedao/design-system'
+import { Button } from '../../../../ui/components/button/index.js'
 import { EnumField } from '../../enum-field/index.js'
 import { Form } from '../../form/index.js'