+++ src/ui/components/object-set-table/subcomponents/table-provider/table-provider.tsx
@@ -31,7 +31,7 @@ const TableProvider = <T extends DataType>({ children, config, tableRef }: Table
   const [state, dispatch] = useReducer(tableReducer<T>, {
     columns: config.columns,
     defaultData: [...config.data],
-    dataFormReferences: createFormReferences(config.data.length, config.columns),
+    dataFormReferences: createFormReferences(config.data.length + 1, config.columns),
     data: config.data,
     allowRowSelection: config.allowRowSelection ?? true,
     showRowNumbers: config.showRowNumbers ?? true,
@@ -41,6 +41,13 @@ const TableProvider = <T extends DataType>({ children, config, tableRef }: Table
     isCellEditMode: false,
     sortState: null,
   })
+
+  useEffect(() => {
+    dispatch({
+      type: 'SET_DATA',
+      payload: config.data,
+    })
+  }, [config.data])
 
   useEffect(() => {
     dispatch({ type: 'SET_DISPATCH', payload: dispatch })