+++ src/ui/components/object-set-table/types.ts
@@ -41,6 +41,20 @@ export interface ObjectSetTableConfig<T> {
    * @default 0
    */
   minRowCount?: number
+
+  /**
+   * A function that is called when one or multiple rows are deleted.
+   *
+   * @param rows The rows that are being deleted.
+   */
+  onDelete?: (rows: T[]) => Promise<void> | void
+
+  /**
+   * A function that is called when a new row is added.
+   *
+   * @param data The data of the new row.
+   */
+  onAdd?: (data: Record<string, unknown>) => Promise<void> | void
 }
 
 export type ColumnType = 'text' | 'number' | 'boolean'
@@ -191,11 +205,6 @@ export interface SortableColumnDef<T = unknown> {
    * @default false
    */
   sortable?: boolean
-
-  /**
-   * The default sort direction, only one column can be sorted at a time.
-   */
-  defaultSortDirection?: SortDirection
 
   /**
    * A function that compares two rows.